move middleware logic to own class
